Refactoring literal finder.
We also permit an Aho-Corasick automaton to be used in more cases.
Generally, if prefix literals are <= 250 bytes, then we use a fully
expanded DFA (a single memory intensive sparse matrix), otherwise we
use a less memory intensive but slower Aho-Corasick DFA.

We still retain an overall limit on the number of size of literals that
are extracted, since the algorithm is trivially exponential. e.g.,
`[0-9]{10}` requires more than 10^10 bytes.